Procore/requirements.txt
fastapi==0.100.0
uvicorn==0.23.2
pydantic==1.10.9
mysql-connector-python==8.0.33
python-dotenv==1.0.0
